Add an `evaluate_extras` function
`evaluate_extras` checks if the requirement should be activated with the given set of extras without evaluating the remaining environment markers, without performing any validation. we need this during dependency resolution because we want to resolve to a consistent set of dependencies that can be installed on any platform (by filtering this resolution by the current environment), but only resolve those optional dependencies that are or can actually required by extras.